ストックボタンをコンポーネント化
