+++ src/features/articles/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useLike } from './useLike';
+export { useStock } from './useStock';
